requirements.txt
python-dotenv
openai
requests
langchain-openai
